+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Added user-defined developer console stats which are displayed on-screen (stats_ commands).
+- Added user-defined developer console stats which are displayed on-screen.
+- Changed preferences file to be deleted when unable to load correctly.
 - Fixed exception not being caught when an issue occurs serialising or deserialising the preferences file.
+- Fixed exception not being caught when reading deserialising an object from the preferences file.
 
 ## [1.0.0] - 2021-08-22
 - Initial major release.